Client -> User 수정 로그인 기능 구현

+++ LetsEatAll/src/main/java/com/letseatall/letseatall/data/Entity/Restaurant.java
@@ -29,5 +29,5 @@ public class Restaurant {
     @Column
     private int category;
     @Column
-    private int fid;       // franchise id
+    private Long fid;       // franchise id
 }
